--- src/uglify/versions.js
@@ -1,5 +1,0 @@
-export default {
-  uglify: require('uglify-es/package.json').version, // eslint-disable-line global-require
-  plugin: require('../../package.json').version, // eslint-disable-line global-require
-};
-
